+++ src/main/java/com/codeup/closetrycorner/services/CategoriesRepo.java
@@ -3,5 +3,5 @@ package com.codeup.closetrycorner.services;
 import com.codeup.closetrycorner.models.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoriesRepo extends CrudRepository<Category, Integer> {
+public interface CategoriesRepo extends CrudRepository<Category, Long> {
 }
