+++ src/main/java/com/codeup/closetrycorner/services/GarmentRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface GarmentRepo extends CrudRepository<Garment, Integer> {
+public interface GarmentRepo extends CrudRepository<Garment, Long> {
     @Query(value = "Select date, image, id from garments where user_id = ?1", nativeQuery = true)
-    List<Garment> findGarmentsByUser(int user_id);
+    List<Garment> findGarmentsByUser(long user_id);
 }
