+++ src/main/java/com/codeup/closetrycorner/services/OutfitsRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface OutfitsRepo extends CrudRepository<Outfit, Integer> {
+public interface OutfitsRepo extends CrudRepository<Outfit, Long> {
     @Query(value = "Select date, image, id from garments where user_id = ?1", nativeQuery = true)
-    List<Outfit> findOutfitsByUser(int user_id);
+    List<Outfit> findOutfitsByUser(long user_id);
 }
